MysqlExtended now wraps execute() with retry
We were getting deadlocks more and more,
after the switch to InnoDB on our more active tables.

After research, we decided that automatic retry of the query was the way to go.
Mysql recommended it (it's in the error message).

This should only matter/happen in the case of PDOException Exceptions
which have the phrase "try restarting transaction" in the message.

This should only happen up to 3 times per query, per run.

In between each attempt there's a very brief pause.

If we do not retry, we just throw the PDOException.

Added logging for now, we may remove that at some point.